+++ lib/packer/postprocessor.rb
@@ -10,6 +10,7 @@ module Packer
     DOCKER_TAG = 'docker-tag'
     VAGRANT = 'vagrant'
     COMPRESS = 'compress'
+    SHELL_LOCAL = 'shell-local'
 
     VALID_POST_PROCESSOR_TYPES = [
       DOCKER_IMPORT,
@@ -17,7 +18,8 @@ module Packer
       DOCKER_SAVE,
       DOCKER_TAG,
       COMPRESS,
-      VAGRANT
+      VAGRANT,
+      SHELL_LOCAL
     ]
 
     class UnrecognizedPostProcessorTypeError < StandardError
@@ -33,6 +35,7 @@ module Packer
         DOCKER_SAVE   => Packer::PostProcessor::DockerSave,
         DOCKER_TAG    => Packer::PostProcessor::DockerTag,
         COMPRESS      => Packer::PostProcessor::Compress,
+        SHELL_LOCAL   => Packer::PostProcessor::ShellLocal,
         VAGRANT       => Packer::PostProcessor::Vagrant
       }.fetch(type).new
     end