+++ lib/packer/postprocessors/all.rb
@@ -2,3 +2,4 @@
 require_relative 'docker'
 require_relative 'vagrant'
 require_relative 'compress'
+require_relative 'shell-local'